Add files via upload
Using the Linthurst dataset, this statistical study analyses the variables affecting the production of airborne biomass in the North Carolina Cape Fear Estuary. The research makes use of a range of regression approaches as well as variable selection methods to pinpoint the essential physicochemical characteristics of the substrate that have an impact on biomass output. The biomass (BIO) is the response variable in the dataset, which consists of 14 predictor variables pertaining to soil properties. The Linthurst project examines how substrate physicochemical characteristics affect the generation of airborne biomass in North Carolina's Cape Fear Estuary. The main objective of the dataset, which consists of several predictor variables describing soil parameters, is to comprehend the relationship with biomass output. There are three sections to the analysis.